+++ addon/services/hosted-service.js
@@ -83,6 +83,24 @@ export default Service.extend(serviceMixin, {
   },
 
   /**
+   * Delete from the service definition
+   */
+  deleteFromDefinition (featureServiceUrl, definition, layerId, portalOpts) {
+    let adminUrl = featureServiceUrl.replace(/\/arcgis\/rest\/services\//i, '/arcgis/rest/admin/services/');
+    let url = `${adminUrl}/deleteFromDefinition?f=json`;
+    if (layerId || layerId === 0) {
+      url = `${adminUrl}/${layerId}/deleteFromDefinition?f=json`;
+    }
+    // since the fs has it's own url we use the requestUrl method
+    return this.requestUrl(url, {
+      method: 'POST',
+      data: {
+        deleteFromDefinition: JSON.stringify(definition)
+      }
+    }, portalOpts);
+  },
+
+  /**
    * Shared logic for POST operations
    */
   _post (urlPath, service, portalOpts) {