+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- `hosted-service:deleteFromDefinition` method
+
 ## 1.4.0
 ### Added
 - `group-service:inviteUsers` method